# docs-site/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PURECORTEX API Gateway",
    "version": "0.7.0"
  },
  "paths": {
...
    "/api/agents/{agent_name}/chat": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Chat With Agent",
        "description": "Chat with a specific protocol agent.\nCurrently returns a placeholder response. Post-TGE, this will route\nthrough the Tri-Brain Orchestrator with agent-specific system prompts.",
        "operationId": "chat_with_agent_api_agents__agent_name__chat_post",
        "parameters": [
          {
            "name": "agent_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessage"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agents/{agent_name}/activity": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Get Agent Activity",
        "description": "Get an agent's recent on-chain and off-chain activity.",
        "operationId": "get_agent_activity_api_agents__agent_name__activity_get",
        "parameters": [
          {
            "name": "agent_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActivityRecord": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "txn_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Txn Id"
          }
        },
        "type": "object",
        "required": [
          "action",
          "detail",
          "timestamp"
        ],
        "title": "ActivityRecord"
      },
      "ActivityResponse": {
        "properties": {
          "agent": {
            "type": "string",
            "title": "Agent"
          },
          "total_actions": {
            "type": "integer",
            "title": "Total Actions"
          },
          "recent_activity": {
            "items": {
              "$ref": "#/components/schemas/ActivityRecord"
            },
            "type": "array",
            "title": "Recent Activity"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "agent",
          "total_actions",
          "recent_activity"
        ],
        "title": "ActivityResponse"
      },
...
      "ChatMessage": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatMessage"
      },
      "ChatResponse": {
        "properties": {
          "agent": {
            "type": "string",
            "title": "Agent"
          },
          "response": {
            "type": "string",
            "title": "Response"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "agent",
          "response",
          "timestamp"
        ],
        "title": "ChatResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
  "servers": [
    {
      "url": "https://purecortex.ai",
      "description": "Production"
    }
  ]
}